requirements.txt
flask
pymongo
google-generativeai
python-dotenv
gunicorn